Load PathUtils and FileVisitor using a BiConsumer instead of calling method on not existing folder
