+++ core/src/main/java/org/testcontainers/utility/PathUtils.java
@@ -4,11 +4,8 @@ import lombok.NonNull;
 import lombok.experimental.UtilityClass;
 
 import java.io.IOException;
-import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 
 /**
  * Filesystem operation utility methods.
@@ -21,21 +18,9 @@ public class PathUtils {
      *
      * @param directory path to the directory to delete.
      */
-    public static void recursiveDeleteDir(final @NonNull Path directory) {
+    public static void recursiveDeleteDir(final @NonNull Path directory, DeleteFileVisitor deleteFileVisitor) {
         try {
-            Files.walkFileTree(directory, new SimpleFileVisitor<Path>() {
-                @Override
-                public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-                    Files.delete(dir);
-                    return FileVisitResult.CONTINUE;
-                }
-
-                @Override
-                public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-                    Files.delete(file);
-                    return FileVisitResult.CONTINUE;
-                }
-            });
+            Files.walkFileTree(directory, deleteFileVisitor);
         } catch (IOException ignored) {
         }
     }
